requirements.txt
arabic-reshaper==3.0.0
# Force a new build

asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.6
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
num2words==0.5.14
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.0.0
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
waitress==3.0.2
webencodings==0.5.1
whitenoise==6.10.0
xhtml2pdf==0.2.17

# Web Push
pywebpush==1.14.0